+++ spec/recipes/default_spec.rb
@@ -4,7 +4,8 @@
 require_relative '../spec_helper'
 
 describe 'spotify::default' do
-  let(:runner) { ChefSpec::SoloRunner.new }
+  let(:platform) { { platform: 'ubuntu', version: '16.04' } }
+  let(:runner) { ChefSpec::SoloRunner.new(platform) }
   let(:chef_run) { runner.converge(described_recipe) }
 
   it 'installs the Spotify app' do