+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Spotify Cookbook CHANGELOG
 Unreleased
 ----------
 - Remove dependency on the apt cookbook
+- Remove dependency on the windows cookbook
 
 v1.0.0 (2016-03-03)
 -------------------